+++ NEWS.md
@@ -1,6 +1,7 @@
 # version 0.8.0 
 
 * get_all --- iterate through the results and get all supported for various functions. supported for yt_search(). prints removed from yt_search()
+* yt_search() for returns a data.frame with video_id when simplify is TRUE
 
 # version 0.7.0